Changed static const doubles to ints in Camera

+++ Camera.h
@@ -46,14 +46,14 @@ private:
 	    //Define the perspective division:
 		double m_fovy, m_aspect, m_znear, m_zfar, m_zoom;
 	    //My variables for camera movement:
-		static const double S_XTRANS_DEFAULT = -14; //0
-		static const double S_YTRANS_DEFAULT = -28; //-20;
-		static const double S_ZTRANS_DEFAULT = -64; //-70
-		static const double S_TILT_DEFAULT = 30;
-		static const double S_YAW_DEFAULT = -21; //0
-		static const double S_CAMERA_INC_DEFAULT = 1; //amount to move on key press
-		static const double S_ROTATE_AMOUNT = 3; //This is left/right rotation
-		static const double S_TILT_AMOUNT = 1; //Up/down
+		static const int S_XTRANS_DEFAULT = -14; //0
+		static const int S_YTRANS_DEFAULT = -28; //-20;
+		static const int S_ZTRANS_DEFAULT = -64; //-70
+		static const int S_TILT_DEFAULT = 30;
+		static const int S_YAW_DEFAULT = -21; //0
+		static const int S_CAMERA_INC_DEFAULT = 1; //amount to move on key press
+		static const int S_ROTATE_AMOUNT = 3; //This is left/right rotation
+		static const int S_TILT_AMOUNT = 1; //Up/down
 		double m_xtrans, m_ytrans, m_ztrans, m_tilt, m_yaw;
 		double m_camera_inc;
 };
